requirements.txt
pdfplumber==0.10.1
PyMuPDF==1.22.5
langdetect==1.0.9
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.6
